+++ opennmt/bin/main.py
@@ -48,7 +48,10 @@ def main():
                       help="Run type.")
   parser.add_argument("--config", required=True, nargs="+",
                       help="List of configuration files.")
-  parser.add_argument("--model_type", default="", choices=list(classes_in_module(catalog)),
+  parser.add_argument("--auto_config", default=False, action="store_true",
+                      help="Enable automatic configuration values.")
+  parser.add_argument("--model_type", default="",
+                      choices=list(classes_in_module(catalog, public_only=True)),
                       help="Model type from the catalog.")
   parser.add_argument("--model", default="",
                       help="Custom model configuration file.")
@@ -147,7 +150,8 @@ def main():
       config,
       seed=args.seed,
       num_devices=args.num_gpus,
-      session_config=session_config)
+      session_config=session_config,
+      auto_config=args.auto_config)
 
   if args.run == "train_and_eval":
     runner.train_and_evaluate()