+++ CHANGELOG.md
@@ -14,6 +14,7 @@ OpenNMT-tf follows [semantic versioning 2.0.0](https://semver.org/). The API cov
 
 ### New features
 
+* `--auto_config` flag to use automatic training configuration values (e.g. optimizer, learning rate, batch size). For compatible models, the automatic values aim to deliver solid performance out of the box.
 * Include all tokenization assets in exported models
 
 ### Fixes and improvements
@@ -21,6 +22,9 @@ OpenNMT-tf follows [semantic versioning 2.0.0](https://semver.org/). The API cov
 * Fix type error during evaluation and inference of FP16 Transformer models
 * Update the [model serving example](https://github.com/OpenNMT/OpenNMT-tf/tree/master/examples/serving) to use a real pretrained model with the TensorFlow Serving 1.11 GPU Docker image
 * Small training speed improvement when the optimizer implements sparse updates
+* Revise some default configuration values:
+  * change `bucket_width` default value to 1 (from 5)
+  * change inference `batch_size` default value to 16 (from 1)
 
 ## [1.9.0](https://github.com/OpenNMT/OpenNMT-tf/releases/tag/v1.9.0) (2018-10-05)